--- springboot-jwt/src/main/java/com/xncoding/jwt/common/util/DateUtil.java
@@ -1,296 +1,0 @@
-/**
- * Copyright (c) 2015-2016, Chill Zhuang 庄骞 (smallchill@163.com).
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.xncoding.jwt.common.util;
-
-import java.sql.Timestamp;
-import java.text.DateFormat;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
-import java.util.Date;
-
-public class DateUtil {
-    private final static SimpleDateFormat sdfYear = new SimpleDateFormat("yyyy");
-
-    private final static SimpleDateFormat sdfDay = new SimpleDateFormat("yyyy-MM-dd");
-
-    private final static SimpleDateFormat sdfDays = new SimpleDateFormat("yyyyMMdd");
-
-    private final static SimpleDateFormat sdfTime = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-
-    private final static SimpleDateFormat sdfmsTime = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS");
-
-    private final static SimpleDateFormat allTime = new SimpleDateFormat("yyyyMMddHHmmss");
-
-    private final static SimpleDateFormat sdfDay_ = new SimpleDateFormat("yyyy/MM/dd");
-
-
-    /**
-     * 获取YYYY格式
-     *
-     * @return
-     */
-    public static String getYear() {
-        return sdfYear.format(new Date());
-    }
-
-    /**
-     * 获取YYYY格式
-     *
-     * @return
-     */
-    public static String getYear(Date date) {
-        return sdfYear.format(date);
-    }
-
-    /**
-     * 获取YYYY-MM-DD格式
-     *
-     * @return
-     */
-    public static String getDay() {
-        return sdfDay.format(new Date());
-    }
-
-    /**
-     * 获取YYYY-MM-DD格式
-     *
-     * @return
-     */
-    public static String getDay(Date date) {
-        return sdfDay.format(date);
-    }
-
-    /**
-     * 获取YYYYMMDD格式
-     *
-     * @return
-     */
-    public static String getDays() {
-        return sdfDays.format(new Date());
-    }
-
-    /**
-     * 获取YYYYMMDD格式
-     *
-     * @return
-     */
-    public static String getDays(Date date) {
-        return sdfDays.format(date);
-    }
-
-    /**
-     * 获取YYYY/MM/DD格式
-     *
-     * @return
-     */
-    public static String getDays_(Date date) {
-        return sdfDay_.format(date);
-    }
-
-    /**
-     * 获取YYYY-MM-DD HH:mm:ss格式
-     *
-     * @return
-     */
-    public static String getTime() {
-        return sdfTime.format(new Date());
-    }
-
-    /**
-     * 获取YYYY-MM-DD HH:mm:ss.SSS格式
-     *
-     * @return
-     */
-    public static String getMsTime() {
-        return sdfmsTime.format(new Date());
-    }
-
-    /**
-     * 获取YYYYMMDDHHmmss格式
-     *
-     * @return
-     */
-    public static String getAllTime() {
-        return allTime.format(new Date());
-    }
-
-    /**
-     * 获取YYYY-MM-DD HH:mm:ss格式
-     *
-     * @return
-     */
-    public static String getTime(Date date) {
-        return sdfTime.format(date);
-    }
-
-    /**
-     * @param s
-     * @param e
-     * @return boolean
-     * @throws
-     * @Title: compareDate
-     * @Description:(日期比较，如果s>=e 返回true 否则返回false)
-     * @author luguosui
-     */
-    public static boolean compareDate(String s, String e) {
-        if (parseDate(s) == null || parseDate(e) == null) {
-            return false;
-        }
-        return parseDate(s).getTime() >= parseDate(e).getTime();
-    }
-
-    /**
-     * 格式化日期
-     *
-     * @return
-     */
-    public static Date parseDate(String date) {
-        try {
-            return sdfDay.parse(date);
-        } catch (ParseException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-    /**
-     * 格式化日期
-     *
-     * @return
-     */
-    public static Date parseTime(String date) {
-        try {
-            return sdfTime.parse(date);
-        } catch (ParseException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-    /**
-     * 格式化日期
-     *
-     * @return
-     */
-    public static Date parse(String date, String pattern) {
-        DateFormat fmt = new SimpleDateFormat(pattern);
-        try {
-            return fmt.parse(date);
-        } catch (ParseException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-    /**
-     * 格式化日期
-     *
-     * @return
-     */
-    public static String format(Date date, String pattern) {
-        DateFormat fmt = new SimpleDateFormat(pattern);
-        return fmt.format(date);
-    }
-
-    /**
-     * 把日期转换为Timestamp
-     *
-     * @param date
-     * @return
-     */
-    public static Timestamp format(Date date) {
-        return new Timestamp(date.getTime());
-    }
-
-    /**
-     * 校验日期是否合法
-     *
-     * @return
-     */
-    public static boolean isValidDate(String s) {
-        try {
-            sdfTime.parse(s);
-            return true;
-        } catch (Exception e) {
-            // 如果throw java.text.ParseException或者NullPointerException，就说明格式不对
-            return false;
-        }
-    }
-
-    /**
-     * 校验日期是否合法
-     *
-     * @return
-     */
-    public static boolean isValidDate(String s, String pattern) {
-        DateFormat fmt = new SimpleDateFormat(pattern);
-        try {
-            fmt.parse(s);
-            return true;
-        } catch (Exception e) {
-            // 如果throw java.text.ParseException或者NullPointerException，就说明格式不对
-            return false;
-        }
-    }
-
-    /**
-     * 获取指定日期偏移指定时间后的时间
-     *
-     * @param date          基准日期
-     * @param calendarField 偏移的粒度大小（小时、天、月等）使用Calendar中的常数
-     * @param offsite       偏移量，正数为向后偏移，负数为向前偏移
-     * @return 偏移后的日期
-     */
-    public static Date offsiteDate(Date date, int calendarField, int offsite) {
-        Calendar cal = Calendar.getInstance();
-        cal.setTime(date);
-        cal.add(calendarField, offsite);
-        return cal.getTime();
-    }
-
-    /**
-     * 返回日期零时
-     *
-     * @param date 目标日期
-     * @return 目标日期零时
-     */
-    public static Date getDateStartTime(Date date) {
-        String str = format(date, "yyyyMMdd") + "000000";
-        try {
-            return allTime.parse(str);
-        } catch (ParseException e) {
-            return null;
-        }
-    }
-
-
-    /**
-     * 返回日期最后时间End
-     *
-     * @param date 目标日期
-     * @return 目标日日期最后时间
-     */
-    public static Date getDateEndTime(Date date) {
-        String str = format(date, "yyyyMMdd") + "235959";
-        try {
-            return allTime.parse(str);
-        } catch (ParseException e) {
-            return null;
-        }
-    }
-
-}
