+++ springboot-jwt/src/main/java/com/xncoding/jwt/service/ManagerInfoService.java
@@ -1,7 +1,8 @@
 package com.xncoding.jwt.service;
 
-import com.xncoding.jwt.dao.entity.ManagerInfo;
-import com.xncoding.jwt.dao.entity.SysRole;
+import com.xncoding.jwt.model.ManagerInfo;
+import com.xncoding.jwt.model.SysRole;
+import com.xncoding.jwt.shiro.ShiroKit;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -18,20 +19,23 @@ public class ManagerInfoService {
 
     /**
      * 通过名称查找用户
+     * 这里我直接写常量，实际生产环境会通过DAO访问数据库
      *
      * @param username
      * @return
      */
     public ManagerInfo findByUsername(String username) {
         ManagerInfo managerInfo = new ManagerInfo();
+        managerInfo.setUsername(username);
         managerInfo.setPids("1,2,3");
         managerInfo.setPidsList(Arrays.asList(1, 2, 3));
         managerInfo.setPnames("第1个,第2个");
         managerInfo.setState(1);
         managerInfo.setCreatedTime(new Date());
         managerInfo.setName("系统管理员");
-        managerInfo.setPassword("4a496ba2a4172c71540fa643ddc8bb7c");
-        managerInfo.setSalt("b4752b4b73034de06afb2db30fe19061");
+        // 随机数
+        managerInfo.setSalt("ef748186673033723bbf4e056f4ec92e");
+        managerInfo.setPassword("da9c3061ae5c0973a3d48e4e721cfbad");
         List<SysRole> roles = new ArrayList<>();
         SysRole role = new SysRole();
         role.setId(1);
