+++ springboot-jwt/src/main/java/com/xncoding/jwt/shiro/ShiroKit.java
@@ -15,11 +15,9 @@
  */
 package com.xncoding.jwt.shiro;
 
-import com.xncoding.jwt.dao.entity.ManagerInfo;
 import org.apache.shiro.SecurityUtils;
 import org.apache.shiro.crypto.SecureRandomNumberGenerator;
 import org.apache.shiro.crypto.hash.SimpleHash;
-import org.apache.shiro.session.Session;
 import org.apache.shiro.subject.Subject;
 
 /**
