--- springboot-jwt/src/test/java/com/xncoding/jwt/ApplicationTests.java
@@ -1,15 +1,0 @@
-package com.xncoding.jwt;
-
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-/**
- * 测试密码加密
- */
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class ApplicationTests {
-
-}
